[OpinionatedWatcher/OpinionatedReconciler] Fix overwriting of finalizers with concurrent apps watching the same kinds (#910)
The patch operation being done currently by the
`operator.OpinionatedWatcher` and `operator.OpinionatedReconciler`
overwrites the finalizers in the list on Add, rather than appending if
there are existing finalizers. As the logic to correctly patch the
finalizers, especially with multiple operators watching the same kinds,
is a little more complex than a single patch request, this PR introduces
the `FinalizerUpdater` interface and an unexported implementation that
is used by both `operator.OpinionatedWatcher` and
`operator.OpinionatedReconciler`.

The `FinalizerUpdater` implementation now handles using resourceVersion
to avoid out-of-date RV issues, and will append to or create the
finalizer list as appropriate on `AddFinalizer`. It also handles a few
retries with a simple backoff and jitter in case multiple operators are
all reacting to the same event simultaneously.

To support the ResourceVersion mismatch get-and-repatch behavior, a
`Get` method had to be added to `k8s.DynamicPatcher`, and `GetInto` to
the `operator.PatchClient` interface. The `GetInto` matches the
signature for the same-name method in `resource.Client`, so this should
avoid being a breaking change for anyone using
`NewOpinionatedWatcher`/`NewOpinionatedReconciler` directly, instead of
through the `simple.App`, provided they are supplying a
`resource.Client`.

This behavior was tested locally on k3d with a binary that had two
operator runners in it each with their own finalizer and WIP finalizer,
watching the same kind.