[codegen] Generate Typed Clients for Resources (#890)
With the addition of custom subresource routes, the generic way of
querying for a subresource using `resource.Client.SubresourceRequest`
cannot be typed in a useful way, and will always consume a generic body
and return generic bytes. To help make it simpler to use custom
subresources, this PR introduces generated, typed clients for each kind
that include methods for each custom subresource that accept strongly
typed parameters and return a typed body response.

This PR also changes the associator functions to use a single interface
rather than adding more functions to `apiserver.NewDefaultInstaller`.