[CLI] Switch YAML encoder for CLI to `sigs.k8s.io/yaml` (#925)
Switch YAML encoder for CLI to `sigs.k8s.io/yaml`. This is a change
needed for larger AppManifest changes and has been broken out to reduce
the size of that PR. `cog`-generated code doesn't include `yaml` tags,
and `gopkg.in/yaml.v3` will use a snake_case name for exported fields if
the `yaml` tag is not present. To use the `json` tags in generated code,
we need to use the kubernetes YAML encoder, which uses the `json` tags
to produce YAML which is identical to JSON generated from the same code
(as with kubernetes the YAML and JSON must have exact parity).

The bulk of the changes in this PR are due to the kubernetes YAML
encoder using a different number of spaces for indentation, and some
field reording, causing all the generated golden YAML files to have a
diff.