[k8s] Allow partial discovery failure for the dynamic patcher (#940)
This PR is a backport of
https://github.com/grafana/grafana-app-sdk/pull/900 onto `lts/v0.24`.

## Copied from [the PR to main]()

The kubernetes `dynamic.DynamicClient`'s `ServerPreferredResources`
method [can return an error alongside
results](https://github.com/kubernetes/client-go/blob/master/discovery/discovery_client.go#L493)
to indicate that some apigroup discoveries failed, but others succeeded.
To handle this, if we have any results, just log the error to the
default logger, and still process the results. If there are no results,
return the error.