[AppManifest/Codegen] Use Full OpenAPI Schemas Section in AppManifest (#929)
This PR introduces a new `v1alpha2` version of the AppManifest which
uses a `schemas` section instead of `schema`. This section is equivalent
to a `components.schemas` section of an OpenAPI v3 document.
`app.VersionSchema` has been updated to have internals which are an
OpenAPI `components.schemas` section, and `AsMap()` has been changed to
`AsOpenAPI3SchemasMap()`. To allow for extracting a CRD-compatible
single schema from the `VersionSchema`, `AsCRDMap(string)` has been
introduced, as well as `AsCRDOpenAPI3(string)`.

To allow for generated manifest JSON/YAML to still include
CRD-compatible schemas, the version of the generated manifest can be set
to `v1alpha1` with the flag `--crdmanifest`. If `--crdmanifest` is
present or set to `true`, the generated AppManifest JSON/YAML will be a
`v1alpha1` AppManifest, and use a CRD-compatible `schema` section.
Otherwise, by default, the new generated AppManifest will be a
`v1alpha2` AppManifest with a `schemas` section.

This change is necessary to fully support custom API servers with
recursive kind definitions, as they need to use references to other
definitions (or themselves), which when converted to CRD-compatible
recursive types, become untyped objects.