[MultiRunner] Fix edge case panic in MultiRunner when a runner would error out exactly when the runner was closing the errors channel (#952)
Very infrequently the `app.MultiRunner` tests would fail with:
```
panic: send on closed channel

goroutine 45 [running]:
github.com/grafana/grafana-app-sdk/app.(*MultiRunner).Run.func2()
        /Users/austinpond/grafana/grafana-app-sdk/app/runner.go:68 
created by github.com/grafana/grafana-app-sdk/app.(*MultiRunner).Run in goroutine 43
        /Users/austinpond/grafana/grafana-app-sdk/app/runner.go:60 
```

This appears to be due to a second (or third, fourth, etc.) runner
erroring out at the same time as `Run()` exits and closes the `errs`
channel. There was a check boolean for if `errs` is already closed, but
no mutex to make sure the state was accurate across the entire execution
block. This PR adds a mutex that is used to make sure the `errsClosed`
boolean is set to true in the same operation as `errs` is closed, and
that writing to `errs` is done in the same operation as checking that
`errsClosed` is false.