Replace the list of finalizers on deletion instead of removing by index (#973)
### What

Change the logic of removing the finalizer on object deletion to instead
replace the list of finalizers completely.

### Why

This fixes an issue with the API server sometimes not being able to
apply the patch, because the provided index to remove at causes an
out-of-bounds error (e.g. if some other client shrunk the list of
finalizers with a PUT / PATCH request in the meantime).

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>