Function to Auto-Retry of Update Conflicts (#972)
A useful feature of the app-sdk would be to have the ability to
auto-retry conflicts, as often an app author will not want to have to
write specify error-based retry handling for their app logic.

This PR introduces a function in the `resource` package called
`UpdateObject`, which does a 409-retryable `Put` request using a client,
and accepts a mutation function rather than an object, in order to be
able to use get-and-put semantics for both the initial and conflict
retries.

This can be used from the stores by calling `Client()` on the store, for
example (updating the status):
```go
updated, err := UpdateObject(ctx, store.Client(), func(obj **v1alpha1.MyObject, isRetry bool) {
    (*obj).Status.MyList = append((*obj).Status.MyList, "to-add")
    return nil
}, resource.UpdateOptions{
    Subresource: resource.SubresourceStatus,
})
```