Improve monitoring and healthcheck setup for operator apps (#978)
### What

This PR updates the app & operator runner to properly propagate metrics
& health checks from the app components and adds a convenience method
for initializing the default logger, which also takes care of
initializing the logger for k8s components from `client-go`.

### Why

To make sure that metrics & healthchecks from app components are
registered correctly and that logging is easier to set up.

---------

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>