[apiserver] Support arbitrary subresources in storage (#956)
Support arbitrary subresources in storage which are declared as part of
the schema of the kind and not a subresource route.

This expands the ability to have `status`-like subresources from only
`status` to any top-level subresource declared in the CUE schema
(anything returned by the kind's `getSubresources()` method).

An example additional subresource has also been added to the
`examples/apiserver` example.

Drive-by bugfix for the go client generator as it didn't properly
differentiate between subresources and always used status for the name.

---------

Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>