[codegen] Add Kubernetes Metadata to Custom Route Responses (#983)
This PR introduces the ability to add/remove kubernetes metadata to
custom route responses, defaulting the route response to having
`metav1.TypeMeta` (`apiVersion` and `kind`), and implementing
`runtime.Object`. Optionally, `ObjectMeta` or `ListMeta` can also be
added as the `metadata` field to the response object, or `TypeMeta` can
be turned off (resulting in the generated object no longer implementing
`runtime.Object`).

This change s controlled via the new `responseMetadata` field in the CUE
custom route object, which is defined as:
```cue
{
    typeMeta: bool | *true
    objectMeta: bool | *false
    listMeta: bool | *false
}
```

On the AppManifest and apiserver side, the OpenAPI generated for the
AppManifest for each response object will include the relevant
kubernetes metadata, but will also include the extension
`x-grafana-app-uses-kubernetes-object-metadata` or
`x-grafana-app-uses-kubernetes-object-metadata` on the `metadata` field
if it uses ObjectMeta or ListMeta respectively. The apiserver installer
uses that extension to convert that field to a reference to the existing
kubernetes ObjectMeta or ListMeta definitions to be more broadly
compatible, rather than using the embedded one.

Resolves https://github.com/grafana/grafana-app-sdk/issues/935

---------

Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>