[APIServer/codegen/AppManifest] Allow Referencing Types in Custom Routes (#1013)
Referencing types (including allowing for recursive types) is currently
allowed in kind schemas, but breaks on custom routes, as only the
response schema is present in the `spec3.PathProps` used in the
AppManifest. This PR introduces the ability to use recrusive
types/references to definitions in custom routes as well, by changing
the way custom routes are parsed from CUE, and adding the referenced
schemas to the manifest.

Custom route openAPI is now generated from cog, and any `$ref`s are
preserved as separate schemas. Custom routes for kinds (subroutes) store
these extra schemas in the `schemas` section of the `v1alpha2`
AppManifest, and custom namespaced/cluster routes store them in a new
`schemas` field in the version's `routes` section.

The way routes and their responses are declared in CUE remains
unchanged, the parser will construct the AppManifest correctly based on
the existing data.

On the API server side, the additional schemas from the kinds and routes
are added to the OpenAPI, and prefixed based on either the kind (for
subresource routes), or operation (namespaced/cluster routes), and the
`$ref` entries in the referencing schemas are updated to point to the
new names (the prefixing is done to avoid conflicts when all these
definitions sit in the same namespace).

As an artifact of switching to using `cog` for route OpenAPI generation,
many custom routes schemas will now have `"additionalProperties": false`
(this should be only in places where this was already implicit).