[CRD Generation] Abide by Structural Schema Constrains for `anyOf`, `allOf`, and `oneOf` (#1018)
Kubernetes CRDs have [particular
restrictions](https://kubernetes.io/blog/2019/06/20/crd-structural-schema/)
on how `anyOf`, `allOf`, and `oneOf` can be used, in that the schema
must be structural, so that the schema must have all parts of the
`anyOf`, `allOf`, or `oneOf`, and the `anyOf`, `oneOf`, or `allOf`
selector is simply used to discriminate on which fields in the schema is
used (see examples in linked document). This is enforced in the
`apiextensions-apiserver` [CRD
validation](https://github.com/kubernetes/apiextensions-apiserver/blob/master/pkg/apiserver/schema/validation.go#L309).

This PR updates `resolveSchema` (used by `app.GetCRDOpenAPISchema`) to
convert the standard OpenAPI schema passed to it into one which abides
by the structural schema restrictions. Tests for
`app.GetCRDOpenAPISchema` have also been added for `anyOf`, `allOf`, and
`oneOf`.

---------

Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>