[APIServer] Enabled Selectable Fields in APIServer (#1014)
Register selectable fields in `installer.AddToScheme` and handle them in
the storage used in `installer.InstallAPIs`. Default to always allowing
`metadata.name`, and `metadata.namespace` if the kind is namespaced.

Tested using the example API server in `examples/apiserver`, and added
to the README for the example how to query with selectable fields.