[AppManifest] Fix panic on `ToManifestData()` call on an AppManifest with no versions (#1087)
## What Changed? Why?

If `ToManifestData` is called on a `v1alpha1.AppManifestSpec` or
`v1alpha2.AppManifestSpec` that has no `PreferredVersion` and no (or
empty) `Versions`, the call will result in a nil pointer panic as it
tries to set `PreferredVersion` to the last element of `Versions`
without checking if `Versions` is non-nil with at least one element.
This PR adds a check that `Versions` has at least one element before
doing this assignment.

### How was it tested?

Added test cases to the conversion tests for this scenario.

### Where did you document your changes?

N/A - bugfix

### Notes to Reviewers