[codegen] Use `generated/manifestdata` for the generated manifest go file(s) (#1100)
## What Changed? Why?

Instead of putting the manifest go file in the base go codegen path
(defaults to `generated`), put it in `<codegenpath>/manifestdata`. This
is done to make the imports that the manifest does of the various kinds
lateral instead of "down" the tree.

If a manifest go file already exists in `generated`, the CLI will
continue to use that file for the manifest data, it will only use the
new `manifestdata` package if there isn't an existing manifest go file
in the root generated directory. This is to preserve legacy behavior for
existing users, who can switch by simply deleting their manifest go file
before running the codegen if they wish.

### How was it tested?

New project, and the codegen tests (plus integration tests which will
run for this PR).

### Where did you document your changes?

Docs that reference the manifest go file's path (tutorial doc) have been
updated.

### Notes to Reviewers