[workflows] Add version comments to golangci-lint and goreleaser commit hashes (#1093)
## What Changed? Why?

Add version comments to golangci-lint and goreleaser commit hashes so
that renovate uses tagged versions when updating instead of trying to
update whenever a new commit is pushed.

The goreleaser workflow was reverted to the hash used before rennovate
updated it to a commit without a version tag.

### How was it tested?

N/A - can't test locally

### Where did you document your changes?

N/A

### Notes to Reviewers