[codegen] Include array item information in the generated go manifest routes (#1125)
## What Changed? Why?

When go manifest files are generated, custom resource routes are
constructed in the go code. When a field in the `SchemaProps` is an
array type, `Items` was not being populated despite existing in the
source OpenAPI being used to generate the file. This has been remedied
so that the `Items` field is now populated.

Additionally, the reference from array items (`$ref`) was not being
prefixed the same way other references were, causing it to point to a
nonexistent entity. This has also been fixed.

### How was it tested?

Locally, using the `apiserver` example, and added codegen tests for
resource routes including an array.

### Where did you document your changes?

N/A - bugfix

### Notes to Reviewers