[codegen] Don't Generate Go Metadata Type (#1094)
## What Changed? Why?

[codegen] Allow GoTypes Jenny to have a configurable set of fields to
exclude from generation when Depth > 0, and exclude the metadata field
from generation using this behavior in the go resource generator. The
`metadata` type is not used and can cause confusion for users, as it is
a merge of the kubernetes object metadata and the custom metadata fields
which are set as annotations in the actual object type. It was
previously generated for compliance with the Thema model, which is no
longer an option when generating code.

### How was it tested?

Tested with `make generate` in a test project, as well as updating the
existing unit tests.

### Where did you document your changes?

Updated the tutorial doc to remove references to the metadata type, as
that appears to be the only place it's referenced.

### Notes to Reviewers